Basic view permission module
This provides a way to alter Django admin to add a view permission to go with
the standard "add/change/delete". This requires a monkeypatch to work around
partial API support and could still use a few tweaks to avoid displaying save
buttons on forms where the user will only get "Permission Denied" errors if they
attempt to save.

This feature also adds a setting (QOURISMAN_STAFF_GLOBAL_VIEW) which can be used
to globally allow staff members access to a list of apps without having to
register each model and assign permissions. You will, of course, still need to
register the model admins in the absence of a global app patcher routine.